refactored some functions into the controller. Updated TODO LIST.
